fix(tests): update spec imports to +page.server after moving loads server-side
PR #423 moved +page.ts loads to +page.server.ts but left the specs importing
from the old ./+page path, breaking typecheck and unit CI jobs.
